requirements.txt
django>=2.2,<2.3
mathparse>=0.1,<0.2
python-dateutil>=2.8,<2.9
sqlalchemy>=1.3,<1.4
pytz~=2021.3
coveralls
flake8
nltk>=3.2,<4.0
nose
pint>=0.8.1
pymongo>=3.3,<4.0
twine
twython
spacy>=3.2,<3.4
sphinx>=3.0,<3.1
sphinx_rtd_theme
pyyaml>=5.3,<5.4
git+git://github.com/gunthercox/chatterbot-corpus@master#egg=chatterbot_corpus
https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.2.0/it_core_news_sm-3.2.0.tar.gz#egg=it_core_news_sm
gunicorn
requests~=2.27.1
whitenoise
python-Levenshtein
